Added new sink for autoscaling purposes.
